--- app/src/main/java/com/keylesspalace/tusky/util/FocalPointEnforcer.kt
@@ -1,192 +1,0 @@
-/* Copyright 2018 Jochem Raat <jchmrt@riseup.net>
- *
- * This file is a part of Tusky.
- *
- * This program is free software; you can redistribute it and/or modify it under the terms of the
- * GNU General Public License as published by the Free Software Foundation; either version 3 of the
- * License, or (at your option) any later version.
- *
- * Tusky is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
- * Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with Tusky; if not,
- * see <http://www.gnu.org/licenses>. */
-
-package com.keylesspalace.tusky.util
-
-import android.graphics.Matrix
-import android.widget.ImageView
-
-import com.keylesspalace.tusky.entity.Attachment.Focus
-import com.squareup.picasso.Callback
-
-/**
- * Enforces the correct cropping of an imageView, to keep the focal point visible at all times.
- *
- * The purpose of this class is to make sure that the focal point information on media
- * attachments are honoured. This class uses the custom matrix option of android ImageView's to
- * customize how the image is cropped into the view.
- *
- * This class also implements the Picasso Callback class, so that it can be passed as a callback
- * to the image loading into the view. This allows the class to set the initial matrix for the
- * correct crop initially. Additionally the [updateFocalPointMatrix] function needs to be called
- * each time the size of the ImageView changes, to ensure that a new crop is calculated.
- *
- * See the explanation of focal points here:
- *    https://github.com/jonom/jquery-focuspoint#1-calculate-your-images-focus-point
- *
- * @param imageView The view for which to enforce the focal points.
- * @param focus The focus object which contains the focal point to enforce.
- */
-class FocalPointEnforcer(private val imageView: ImageView, private val focus: Focus) : Callback {
-    // This companion object defines the helper functions used:
-    companion object {
-        /**
-         * Create the appropriate matrix for the given parameters.
-         *
-         * How it works is using the following steps:
-         *   - First we determine if the image is too wide or too tall for the view size. If it is
-         *   too wide, we need to crop it horizontally and scale the height to fit the view
-         *   exactly. If it is too tall we need to crop vertically and scale the width to fit the
-         *   view exactly.
-         *   - Then we determine what translation is needed to get the focal point in view. We
-         *   prefer to get the focal point at the center of the preview. However if that would
-         *   result in some part of the preview being empty, we instead align the image so that it
-         *   fills the view, but still the focal point is always in view.
-         *
-         * @param viewWidth The width of the imageView.
-         * @param viewHeight The height of the imageView
-         * @param imageWidth The width of the actual image
-         * @param imageHeight The height of the actual image
-         *
-         * @return The matrix which correctly crops the image
-         */
-        fun createFocalMatrix(viewWidth: Float,
-                              viewHeight: Float,
-                              imageWidth: Float,
-                              imageHeight: Float,
-                              focus: Focus): Matrix {
-            val mat = Matrix()
-
-            // calculate scaling:
-            val scale = calculateScaling(viewWidth, viewHeight, imageWidth, imageHeight)
-            mat.preScale(scale, scale)
-
-            // calculate offsets:
-            var top = 0f
-            var left = 0f
-            if (isVerticalCrop(viewWidth, viewHeight, imageWidth, imageHeight)) {
-                top = focalOffset(viewHeight, imageHeight, scale, focalYToCoordinate(focus.y))
-            } else { // horizontal crop
-                left = focalOffset(viewWidth, imageWidth, scale, focalXToCoordinate(focus.x))
-            }
-
-            mat.postTranslate(left, top)
-
-            return mat
-        }
-
-        /**
-         * Calculate the scaling of the image needed to make it fill the screen.
-         *
-         * The scaling used depends on if we need a vertical of horizontal crop.
-         */
-        fun calculateScaling(viewWidth: Float, viewHeight: Float,
-                             imageWidth: Float, imageHeight: Float): Float {
-            if (isVerticalCrop(viewWidth, viewHeight, imageWidth, imageHeight)) {
-                return viewWidth / imageWidth
-            } else {     // horizontal crop:
-                return viewHeight / imageHeight
-            }
-        }
-
-        /**
-         * Return true if we need a vertical crop, false for a horizontal crop.
-         */
-        fun isVerticalCrop(viewWidth: Float, viewHeight: Float,
-                           imageWidth: Float, imageHeight: Float): Boolean {
-            val viewRatio = viewWidth / viewHeight
-            val imageRatio = imageWidth / imageHeight
-
-            return viewRatio > imageRatio
-        }
-
-        /**
-         * Transform the focal x component to the corresponding coordinate on the image.
-         *
-         * This means that we go from a representation where the left side of the image is -1 and
-         * the right side +1, to a representation with the left side being 0 and the right side
-         * being +1.
-         */
-        fun focalXToCoordinate(x: Float): Float {
-            return (x + 1) / 2
-        }
-
-        /**
-         * Transform the focal y component to the corresponding coordinate on the image.
-         *
-         * This means that we go from a representation where the bottom side of the image is -1 and
-         * the top side +1, to a representation with the top side being 0 and the bottom side
-         * being +1.
-         */
-        fun focalYToCoordinate(y: Float): Float {
-            return (-y + 1) / 2
-        }
-
-        /**
-         * Calculate the relative offset needed to focus on the focal point in one direction.
-         *
-         * This method works for both the vertical and horizontal crops. It simply calculates
-         * what offset to take based on the proportions between the scaled image and the view
-         * available. It also makes sure to always fill the bounds of the view completely with
-         * the image. So it won't put the very edge of the image in center, because that would
-         * leave part of the view empty.
-         */
-        fun focalOffset(view: Float, image: Float,
-                        scale: Float, focal: Float): Float {
-            // The fraction of the image that will be in view:
-            val inView = view / (scale * image)
-            var offset = 0f
-
-            // These values indicate the maximum and minimum focal parameter possible while still
-            // keeping the entire view filled with the image:
-            val maxFocal = 1 - inView / 2
-            val minFocal = inView / 2
-
-            if (focal > maxFocal) {
-                offset = -((2 - inView) / 2) * image * scale + view * 0.5f
-            } else if (focal > minFocal) {
-                offset = -focal * image * scale + view * 0.5f
-            }
-
-            return offset
-        }
-    }
-
-    // Called when the Callback succeeds, initializes the focal point crop on the image
-    override fun onSuccess() {
-        updateFocalPointMatrix()
-    }
-
-    // We do not handle the error here, instead it will be handled higher up the call chain.
-    override fun onError() {
-    }
-
-    /**
-     * Update the focal point matrix.
-     *
-     * This updates the focal point matrix based on the new size of the view and
-     * applies this new matrix to the view.
-     */
-    fun updateFocalPointMatrix() {
-        val matrix = createFocalMatrix(imageView.width.toFloat(),
-                imageView.height.toFloat(),
-                imageView.drawable.intrinsicWidth.toFloat(),
-                imageView.drawable.intrinsicHeight.toFloat(),
-                focus)
-
-        imageView.imageMatrix = matrix
-    }
-
-}
